databaseAttempt2

+++ app/src/main/java/com/RuffinApps/johnnie/oldtimeradio/PlayedList.java
@@ -1,9 +1,16 @@
 package com.RuffinApps.johnnie.oldtimeradio;
 
+import android.content.ContentValues;
 import android.content.Context;
+import android.database.Cursor;
 import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.util.Log;
+
+import java.sql.Struct;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Created by MOTAH on 12/19/2016.
@@ -11,22 +18,30 @@ import android.database.sqlite.SQLiteOpenHelper;
 
 public class PlayedList extends SQLiteOpenHelper {
 
+    String TAG = "PlayList";
 
+    private static PlayedList mDbHelper;
     private SQLiteDatabase database;
     private SQLiteOpenHelper dbHelper;
 
     private static final int DATABASE_VERSION = 1;
     private static final String DATABASE_NAME = "played.db";
     private static final String PLAYED_TABLE_NAME = "played";
-    private static final String KEY_WORD = "show_id";
-    private static final String KEY_DEFINITION = "radio_show_title";
+    private static final String SHOW_ID = "show_id";
+    private static final String SHOW_TITLE = "radio_show_title";
     private static final String PLAYED_TABLE_CREATE =
             "CREATE TABLE " + PLAYED_TABLE_NAME + " (" +
-                    KEY_WORD + "TEXT, " +
-                    KEY_DEFINITION + " TEXT);";
+                    SHOW_ID + "TEXT, " +
+                    SHOW_TITLE + " TEXT);";
 
-    PlayedList(Context context) {
-        super(context, DATABASE_NAME, null, DATABASE_VERSION);
+    public static synchronized PlayedList getInstance(Context context) {
+        // Use the application context, which will ensure that you
+        // don't accidentally leak an Activity's context.
+
+        if (mDbHelper == null) {
+            mDbHelper = new PlayedList(context.getApplicationContext());
+        }
+        return mDbHelper;
     }
 
     @Override
@@ -36,31 +51,96 @@ public class PlayedList extends SQLiteOpenHelper {
 
     public void onUpgrade(SQLiteDatabase db, int oldVer, int newVer)
     {
-        db.execSQL("DROP TABLE IF EXISTS " + PLAYED_TABLE_NAME);
-        onCreate(db);
+        if (oldVer != newVer) {
+            db.execSQL("DROP TABLE IF EXISTS " + PLAYED_TABLE_NAME);
+            onCreate(db);
+        }
     }
 
-    public void add(String title)
+   /* public List<String> getPlayedList()
     {
+        List<String> playedList = new ArrayList<>();
 
+        String PLAYED_LIST_SELECT_QUERY = "SELECT * FROM " + PLAYED_TABLE_NAME;
+
+        SQLiteDatabase db = getReadableDatabase();
+
+        Cursor cursor = db.rawQuery(PLAYED_LIST_SELECT_QUERY, null);
+
+        try {
+            if (cursor.moveToFirst()) {
+                do {
+
+                    playData.showId = cursor.getString(cursor.getColumnIndex(NAME));
+                    userData.college = cursor.getString(cursor.getColumnIndex(COLLEGE));
+                    userData.place = cursor.getString(cursor.getColumnIndex(PLACE));
+                    userData.user_id = cursor.getString(cursor.getColumnIndex(USER_ID));
+                    userData.number = cursor.getString(cursor.getColumnIndex(NUMBER));
+
+
+                    usersdetail.add(userData);
+
+                } while (cursor.moveToNext());
+            }
+        } catch (Exception e) {
+            Log.d(TAG, "Error while trying to get posts from database");
+        } finally {
+            if (cursor != null && !cursor.isClosed()) {
+                cursor.close();
+            }
+        }
+    }*/
+
+    public void add(String showId, String title)
+    {
+        store(showId, title);
     }
 
-    public void remove(String title)
+    public void remove(String showId, String title)
     {
-
+       delete(showId, title);
     }
 
-    private boolean store(String title)
-    {
-        return true;
+    private PlayedList(Context context) {
+        super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
 
-    private boolean delete(String title)
+    private boolean store(String showId, String title)
     {
-        System.out.println("Comment deleted with title: " + title);
-        database.delete(KEY_WORD, KEY_DEFINITION
+        boolean successfulStore = true;
+
+        // Make database writable.
+        open();
+
+        database.beginTransaction();
+
+        try {
+            ContentValues values = new ContentValues();
+            values.put(SHOW_ID, showId);
+            values.put(SHOW_TITLE, title);
+            database.insertOrThrow(PLAYED_TABLE_NAME, null, values);
+            database.setTransactionSuccessful();
+        } catch (SQLException e) {
+            e.printStackTrace();
+            Log.d(TAG, "Error while attempting add to Played database.");
+            successfulStore = false;
+        }
+        finally {
+            database.endTransaction();
+            close();
+        }
+        return successfulStore;
+    }
+
+    private boolean delete(String showId, String title)
+    {
+        boolean deleteSuccessful = true;
+
+        Log.d(TAG, "Comment deleted with title: " + title);
+        database.delete(showId, title
                 + " = " + title, null);
-        return true;
+        // TODO: Input query to verify that the entry has been deleted.
+        return deleteSuccessful;
     }
 
     public void open() throws SQLException {
@@ -70,6 +150,4 @@ public class PlayedList extends SQLiteOpenHelper {
     public void close() {
         dbHelper.close();
     }
-
-
 }
